+++ skeleton/SdogDB.cpp
@@ -90,7 +90,7 @@ void SdogDB::insertAirSigmet(const std::vector<std::string>& interior, const std
 	sqlite3_step(insert0Stmt);
 	sqlite3_finalize(insert0Stmt);
 
-	int airSigmetID = sqlite3_last_insert_rowid(db);
+	int airSigmetID = (int)sqlite3_last_insert_rowid(db);
 
 	// Insert the geometry of the AirSigmet into the DB
 	char* sqlInsert1 = "INSERT INTO airSigmetBounds(airSigmetID, number, latRad, longRad) VALUES (@ID, @NUM, @LAT, @LNG)";
@@ -124,7 +124,7 @@ void SdogDB::insertAirSigmet(const std::vector<std::string>& interior, const std
 
 		if (sqlite3_step(selectStmt) == SQLITE_DONE) {
 			insertCell(code);
-			cellID = sqlite3_last_insert_rowid(db);
+			cellID = (int)sqlite3_last_insert_rowid(db);
 		}
 		else {
 			cellID = sqlite3_column_int(selectStmt, 0);
@@ -152,7 +152,7 @@ void SdogDB::insertAirSigmet(const std::vector<std::string>& interior, const std
 
 		if (sqlite3_step(selectStmt) == SQLITE_DONE) {
 			insertCell(code);
-			cellID = sqlite3_last_insert_rowid(db);
+			cellID = (int)sqlite3_last_insert_rowid(db);
 		}
 		else {
 			cellID = sqlite3_column_int(selectStmt, 0);
@@ -312,8 +312,8 @@ void SdogDB::getWindCells(std::vector<std::pair<std::string, glm::vec2>>& out) {
 
 		std::string code = std::string(reinterpret_cast<const char*>(sqlite3_column_text(stmt, 0)));
 		glm::vec3 vec;
-		vec.x = sqlite3_column_double(stmt, 1);
-		vec.y = sqlite3_column_double(stmt, 2);
+		vec.x = (float)sqlite3_column_double(stmt, 1);
+		vec.y = (float)sqlite3_column_double(stmt, 2);
 
 		out.push_back(std::pair<std::string, glm::vec2>(code, vec));
 	}
